backend/requirements.txt
fastapi
sqlmodel
pymysql 
python-dotenv
boto3
workos
uvicorn
email-validator
cryptography
paramiko
celery
Jinja2
redis
ruff
